msm: mdss: fb: wait for idle only on last fb release
There could be multiple references to the framebuffer, when
releasing the last one the panel should be blanked. When blanking the
panel we should make sure that panel update has finished (wait for panel
idle), however this wait should be done only if it will lead to a
change in the panel. Avoid this wait if there are more than one ref
counts which means that panel won't be blanked.

Change-Id: Iffe9a3a1291a5c326107d1127c4e2de41b3284e2
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>